+++ sdk-core/src/main/java/co/reachfive/identity/sdk/core/models/AuthToken.kt
@@ -1,6 +1,7 @@
 package co.reachfive.identity.sdk.core.models
 
 import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
 import kotlinx.parcelize.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
 
@@ -8,10 +9,13 @@ import kotlinx.parcelize.Parcelize
 data class AuthToken(
     // If the `openid` scope is not provided, the `idToken` is not returned
     val idToken: String? = null,
-    val accessToken: String,
-    val refreshToken: String?,
-    val tokenType: String?,
-    val expiresIn: Int?,
+    val accessToken: String? = null,
+    val refreshToken: String? = null,
+    val tokenType: String? = null,
+    val expiresIn: Int? = null,
+    @SerializedName("token")
+    val stepUpToken: String? = null,
+    val amr: List<String>? = null,
     // The `user` field is optional because if the `openid` scope is not provided, the `user` is not retrieved
     val user: OpenIdUser? = null
 ) : Parcelable {