+++ sdk-core/src/main/java/co/reachfive/identity/sdk/core/models/responses/AuthenticationToken.kt
@@ -1,9 +1,12 @@
 package co.reachfive.identity.sdk.core.models.responses
 
 import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class AuthenticationToken(
-    val tkn: String
+    val tkn: String,
+    @SerializedName("mfa_required")
+    val mfaRequired: Boolean? = null
 ) : Parcelable